% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightfly.R
\docType{class}
\name{Finance-class}
\alias{Finance-class}
\alias{Finance}
\title{A base Reference Class for finance.}
\description{
A base Reference Class for finance.
}
\section{Fields}{

\describe{
\item{\code{app_id}}{app id.}

\item{\code{secret_key}}{app secret key.}
}}


