% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightfly.R
\docType{class}
\name{Fund-class}
\alias{Fund-class}
\alias{Fund}
\title{A Fund Reference Class.}
\description{
A Fund Reference Class.
}
\section{Fields}{

\describe{
\item{\code{app_id}}{app id.}

\item{\code{secret_key}}{app secret key.}
}}


