% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightfly.R
\docType{class}
\name{Stock-class}
\alias{Stock-class}
\alias{Stock}
\title{A Stock Reference Class.}
\description{
A Stock Reference Class.
}
\section{Fields}{

\describe{
\item{\code{app_id}}{app id.}

\item{\code{secret_key}}{app secret key.
get HS300
get daily history}

\item{\code{symbol}}{symbol.}

\item{\code{date_from}}{date from.}

\item{\code{date_to}}{date to.
get real time stock}

\item{\code{symbol}}{symbol.}
}}


